some refactoring and add GrantHandler interface and implementations to separate logic pertaining to a specific grant

+++ src/main/kotlin/no/nav/security/mock/oauth2/OAuth2Response.kt
@@ -30,7 +30,7 @@ data class OAuth2TokenResponse(
     @JsonProperty("refresh_token")
     val refreshToken: String? = null,
     @JsonProperty("expires_in")
-    val expiresIn: Long = 0,
+    val expiresIn: Int = 0,
     @JsonProperty("scope")
     val scope: String? = null
 )
